feature/ticket-send-improvements

+++ src/lib/send.ts
@@ -2,14 +2,10 @@ import fs from "fs";
 import { generateInvoice } from "./invoice";
 import { generateTickets } from "./ticket";
 import prisma from "./prisma";
-import { ShippingFactory } from "../store/factories/shipping/ShippingFactory";
-import { DownloadShipping } from "../store/factories/shipping/DownloadShipping";
+import { ShippingFactory, ShippingType } from "../store/factories/shipping/ShippingFactory";
 import { getEmailTransporter } from "./email";
 import ejs from "ejs";
-import {
-    PaymentFactory,
-    PaymentType,
-} from "../store/factories/payment/PaymentFactory";
+import { PaymentFactory, PaymentType } from "../store/factories/payment/PaymentFactory";
 import { getStaticAssetFile } from "../constants/serverUtil";
 import { totalTicketAmount } from "../constants/util";
 import { IOrder } from "../store/reducers/orderReducer";
@@ -27,34 +23,39 @@ export const send = async (orderId) => {
                 tickets: true,
                 paymentResult: true,
                 paymentType: true,
+                invoiceSent: true
             },
         });
 
+       let attachments = [];
+       let invoicePath = undefined;
         // generate invoice
-        const invoicePath = await generateInvoice(
-            getStaticAssetFile("invoice/template.html", "utf-8"),
-            orderId
-        );
+       if (!order.invoiceSent) {
+           invoicePath = await generateInvoice(
+               getStaticAssetFile("invoice/template.html", "utf-8"),
+               orderId
+           );
 
-        const message: any = {
-            from: process.env.EMAIL_SENDER,
-            to: order.user.email,
-            subject: "Your ticket order!",
-            html: null,
-            attachments: [
-                {
-                    filename: "Invoice.pdf",
-                    path: invoicePath,
-                    contentType: "application/pdf",
-                },
-            ],
-        };
+           await prisma.order.update({
+               where: {
+                   id: orderId
+               },
+               data: {
+                   invoiceSent: true
+               }
+           });
+
+           attachments.push({
+               filename: "Invoice.pdf",
+               path: invoicePath,
+               contentType: "application/pdf",
+           });
+       }
 
         // generate tickets
         const shipping = ShippingFactory.getShippingInstance(
             JSON.parse(order.shipping)
         );
-        // TODO: Replace by factory
         const ticketsAlreadySent =
             totalTicketAmount(JSON.parse(order.order) as IOrder) <=
             order.tickets.length;
@@ -65,7 +66,7 @@ export const send = async (orderId) => {
             })?.paymentResultValid(order.paymentResult) ?? false;
         let containsTickets = undefined;
         if (
-            shipping instanceof DownloadShipping &&
+            shipping.shippingData.type === ShippingType.Download &&
             !ticketsAlreadySent &&
             payed
         ) {
@@ -74,7 +75,7 @@ export const send = async (orderId) => {
                 orderId
             );
             tickets.forEach((ticket, i) => {
-                message.attachments.push({
+                attachments.push({
                     filename: `Ticket${i + 1}.pdf`,
                     content: ticket,
                     contentType: "application/pdf",
@@ -82,11 +83,26 @@ export const send = async (orderId) => {
             });
             containsTickets = true;
         }
+
+        if (attachments.length === 0) {
+            resolve();
+            return;
+        }
+
+        const message: any = {
+            from: process.env.EMAIL_SENDER,
+            to: order.user.email,
+            subject: "Your ticket order!",
+            html: null,
+            attachments
+        };
+
         message.html = ejs.render(
             getStaticAssetFile("email/template.html", "utf-8"),
             {
                 customerName: order.user.firstName + " " + order.user.lastName,
                 containsTickets: containsTickets,
+                containsInvoice: invoicePath === undefined ? undefined : true
             }
         );
 
@@ -95,7 +111,8 @@ export const send = async (orderId) => {
                 reject(error);
                 return;
             }
-            fs.unlinkSync(invoicePath);
+            if (invoicePath !== undefined)
+                fs.unlinkSync(invoicePath);
             resolve();
         });
     });
